refactor(carve S6): reduce this slice to the iteration work, without decomposition
Decomposition stays on the development branch as experimental rather than landing
on main: it presupposes one way of working — an issue split into a sub-issue graph
with a human approval gate on the proposed plan — and main should not presuppose
that. The iteration work in this slice does not: it applies to any task that
opened a PR, however that task came to exist.

Removes the 20 decomposition files (planner, caps, flow, store, types, writeback,
render, the plan-command and plan-revise handlers, and their tests). What remains
is the 12 files this slice is now about: the iteration heartbeat and its sweep, the
reply claim, clarify-resume, and the failure reply.

Two helpers in the deleted label module were NOT decomposition and had to survive:
`DEFAULT_LABEL_FILTER`, which the project-mapping table treats as the default when
a project sets no `label_filter` and the rollup renders in operator-facing copy,
and `hasHelpLabel`, the one-time explainer label. They now live in
`trigger-label.ts`, named for what they are. `triggerLabelVariants` was NOT kept —
it existed to return the `:decompose` and `:auto` variants, so with those labels
gone it would only ever return the bare base.

Tests came across with the module, plus one pinning `DEFAULT_LABEL_FILTER`'s value
directly. That constant is load-bearing in a way its size hides: changing it
silently stops every project that never set a filter explicitly.

Slice size drops from 8538 lines to 1918. The PR title and description need to
change to match, since this is no longer a decomposition slice.
